Some updates to make lib more friendly to newcomers
