+++ .gitignore
@@ -11,4 +11,5 @@
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
-.idea
+# IDE
+/.idea/
