+++ go.mod
@@ -1,0 +1,6 @@
+module github.com/chippydip/go-sc2ai
+
+require (
+	github.com/gogo/protobuf v1.2.0
+	github.com/gorilla/websocket v1.4.0
+)
